Changed buttons with GameObjects

+++ LowPop/Assets/Scripts/MoveImage.cs
@@ -1,13 +1,13 @@
 ﻿using UnityEngine;
 
 public class MoveImage : MonoBehaviour {
-    private int x;
-    private int y;
+    private float x;
+    private float y;
     private float time;
 
 	void Start () {
-        x = Random.Range(2, 5);
-        y = Random.Range(2, 5);
+        x = Random.Range(2f, 5f);
+        y = Random.Range(2f, 5f);
         time = 10f;
 
         if (transform.position.y > 0)
